Standard landing page
This looks like enough of a landing page to get started with.  It
provides just enough place for us to position the two services together
and provide some place through which we can show the interface.

This looks like a fair compensation in time for reshaping an LDES feed
dump during the day.